licenta2/requirements.txt
Flask~=3.0.3
python-dotenv
psycopg2-binary